Move ProcessChecksRunInCoreAgent to leaf package to avoid import bloat
Move the helper from pkg/process/util (which transitively pulls in
go4.org/netipx) to a new zero-dependency leaf package
pkg/process/util/coreagent. This avoids adding unnecessary imports to
binaries like cluster-agent-cloudfoundry and otel-agent.
